% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bg_pca.R
\name{bg_pca}
\alias{bg_pca}
\title{bg_pca}
\usage{
bg_pca(
  pca,
  axes = c(1, 2),
  metadata = NULL,
  site.scaling = 1,
  species.scaling = 2,
  goodness.axis = 2,
  goodness.thresh = 0.3,
  main.group = NULL,
  second.group = NULL,
  nudge.x = NULL,
  nudge.y = NULL,
  scale.fill = NULL,
  scale.shape = NULL,
  scale.colour = NULL,
  labels = NULL,
  print.sp.thresh = FALSE,
  ext.plot.scale = 5,
  add.stat1 = TRUE,
  stat1 = "chull",
  add.stat2 = FALSE,
  stat2 = NULL,
  add.dens = TRUE,
  add.centroids = FALSE,
  point.size = 2.5,
  c.size = 5,
  p.alpha = 1,
  font.size = 18/.pt,
  axis.size = 22,
  axis.text = 20,
  conf.level = 0.95,
  ysites = NULL,
  xsites = NULL,
  ysp = NULL,
  xsp = NULL,
  add.labels = TRUE
)
}
\arguments{
\item{pca}{a vegan pca output}

\item{axes}{numeric vector;the selected pca axes to plot}

\item{metadata}{a table containing the metadata mandatory to define the
main and second groups}

\item{site.scaling}{numeric; scaling for the first plot -  defaut is
scaling 1}

\item{species.scaling}{numeric; scaling for the second plot - defaut is
scaling 2}

\item{goodness.axis}{numeric; the number of axis used to calculate the goodness
of fit of the species}

\item{goodness.thresh}{numeric; goodness of fit threshold for selecting species}

\item{main.group}{factor; defines the grouping variable used to calculate the
densities and to define the color and fill aesthethics, it cannot be empty}

\item{second.group}{factor; defines a second level of grouping that, if present,
will be highlighted with different shape}

\item{nudge.x}{numeric vector; defines the gap between the labels and the group
centroid in the x axis, it must have the same length as the grouping variables
defined in main.group}

\item{nudge.y}{numeric vector; defines the gap between the labels and the group
centroid in the y axis, it must have the same length as the grouping variables
defined in main.group}

\item{scale.fill}{string; contains the palette for fill aesthetics}

\item{scale.shape}{numeric vector; contains the shape types for shape aesthetics}

\item{scale.colour}{string; contains the palette for colour aesthetics}

\item{labels}{factor; factor in metadata containing the label of each level of
the grouping variables in main.group or second.group}

\item{print.sp.thresh}{logical; print the goodness of fit threshold for
selecting species' vectors}

\item{ext.plot.scale}{numeric; scale for density plots in proportion to other
plots}

\item{add.stat1}{logical; add either a convex hull or an ellipse to each group
on the first plot}

\item{stat1}{character; choose between a convex hull "chull" or an ellipse
"ellipse" for the first plot}

\item{add.stat2}{logical; add either a convex hull or an ellipse to each group
on the second plot}

\item{stat2}{character; choose between a convex hull "chull" or an ellipse
"ellipse" for the second plot}

\item{add.dens}{logical; add the density plots}

\item{add.centroids}{logical; add centroids for each group}

\item{point.size}{numeric; define sites' point size}

\item{c.size}{numeric; define centroids' point size}

\item{p.alpha}{numeric;  define the alpha for sites' points}

\item{font.size}{character;  defines the font size for geom_label using
the syntax "18/.pt"}

\item{axis.size}{numeric; defines the font size for the axes titles}

\item{axis.text}{numeric; defines the font size fot the axes text}

\item{conf.level}{numeric; defines the C.I. of the ellipse}

\item{ysites}{numeric vector; force y axis limits for the first plot}

\item{xsites}{numeric vector; force x axis limits for the first plot}

\item{ysp}{numeric vector; force y axis limits for the second plot}

\item{xsp}{numeric vector; force x axis limits for the second plot}

\item{add.labels}{logical; add geom_label for plotting the group names}
}
\value{
returns a grid.arrange with two plots with the species and sites
scaling biplots
}
\description{
This function is built on pca output of vegan rda function and
uses Aurelien Boye's code as a basis for creating pretty PCAs but
with a lot of new cosmetic options
}
\examples{
not today, sorry...
}
