% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgplot-package.R
\docType{package}
\name{bgplot-package}
\alias{bgplot}
\alias{bgplot-package}
\title{bgplot: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/victorleitejardim/bgplot}
  \item Report bugs at \url{https://github.com/victorleitejardim/bgplot/issues}
}

}
\author{
\strong{Maintainer}: Victor Leite Jardim \email{vleitejardim@gmail.com} (\href{https://orcid.org/0000-0001-5192-6485}{ORCID}) [copyright holder]

}
\keyword{internal}
